Add files via upload
Based on several medical and demographic characteristics, this machine learning experiment seeks to estimate the chance of mortality from COVID-19 for a specific patient. To train a prediction model, the study uses historical data on COVID-19 patients, including their age, gender, pre-existing medical disorders, and symptoms. This data is used by the model to detect patterns and links between different characteristics and patient outcomes, allowing it to generate accurate predictions about the chance of mortality for new patients. The project's ultimate purpose is to offer healthcare practitioners with a tool that will allow them to make better educated treatment and resource allocation decisions.